+++ src/base/events.h
@@ -72,6 +72,11 @@ typedef enum _event_type_t {
    */
   EVT_WHEEL,
   /**
+   * @const EVT_WHEEL_BEFORE_CHILDREN
+   * 鼠标滚轮事件名，在子控件处理之前触发(key_event_t)。
+   */
+  EVT_WHEEL_BEFORE_CHILDREN,
+  /**
    * @const EVT_POINTER_DOWN_ABORT
    * 取消前一个指针按下事件名(pointer_event_t)。
    */